Proofreading edits
Assuming the second "role" in line 63 should be "ID"?

I think the story in line 79 needs to be fleshed out. Is this something that actually happened?

Should "organisation1" in line 126 be "resourceorganisation1" to match line 116?

This page refers to the user as "it" in several places. This is confusing, since normally we refer to people as "he or she" or "they" or something similar. Might be worth recasting those sentences.

I don't think the last sentence on the page is finished.